% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{clean_time}
\alias{clean_time}
\title{Clean time column}
\usage{
clean_time(.data, .time_col)
}
\arguments{
\item{.data}{Float : Dataframe}

\item{.time_col}{String : Name of the time column from the plate reader experiment}
}
\value{
The dataframe with the formated time column as elapsed hours
}
\description{
Clean time column
}
