% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{format_wellplate_long}
\alias{format_wellplate_long}
\title{Format into long format}
\usage{
format_wellplate_long(.data, wells = matches(regex("^[A-Za-z]{1}\\\\d{1,2}")))
}
\arguments{
\item{.data}{Float : Dataframe}

\item{wells}{Tidyselect : Tidyselect matching all the wells columns from your plate reader experiment}
}
\value{
The dataframe in long/tidy format
}
\description{
Format into long format
}
