% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{richard}
\alias{richard}
\title{Modified Richard Equation}
\usage{
richard(t, p_max, p_min, r_max, s)
}
\arguments{
\item{t}{Float : Time}

\item{p_max}{Float : Population Maximum (Upper Asymptote)}

\item{p_min}{Float : Population Minimum (Lower Asymptote)}

\item{s}{Float : Shift (Time at which r_max occurs)}

\item{r_min}{Float : Maximum Growth/Death rate (positive for growth|negative fo death)}
}
\value{
The population at time \code{t} Given by \deqn{P(t) = p_{min} + \frac{p_{max}-p_{min}}{1 + e^{4r_{max}.(t-s)/p_{min}- p_{max}}}}
}
\description{
Modified Richard Equation
}
