% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{RichardModel}
\alias{RichardModel}
\title{Richard equation}
\value{
The population at time \code{t} Given by \deqn{P(t) = p_{min} + \frac{p_{max}-p_{min}}{1 + e^{4r_{max}.(t-s)/p_{min}- p_{max}}}}
}
\description{
Return the population at time t
}
\section{Super class}{
\code{\link[CuRveR:ModelABC]{CuRveR::ModelABC}} -> \code{RichardModel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-equation}{\code{RichardModel$equation()}}
\item \href{#method-new}{\code{RichardModel$new()}}
\item \href{#method-clone}{\code{RichardModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CuRveR" data-topic="ModelABC" data-id="optimize">}\href{../../CuRveR/html/ModelABC.html#method-optimize}{\code{CuRveR::ModelABC$optimize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CuRveR" data-topic="ModelABC" data-id="predict">}\href{../../CuRveR/html/ModelABC.html#method-predict}{\code{CuRveR::ModelABC$predict()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-equation"></a>}}
\if{latex}{\out{\hypertarget{method-equation}{}}}
\subsection{Method \code{equation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RichardModel$equation(x, p_max, p_min, r_max, s)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p_max}}{Float : Population Maximum (Upper Asymptote)}

\item{\code{p_min}}{Float : Population Minimum (Lower Asymptote)}

\item{\code{s}}{Float : Shift (Time at which r_max occurs)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RichardModel$new(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RichardModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
